requirements.txt
asgiref
cycler
Django
future
kiwisolver
matplotlib
opencv-python
Pillow
pyparsing
python-dateutil
pytz
six
sqlparse
torch
torchvision
